+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Init script now removes template author from pyproject.toml (developers no longer inherit template author info)
+
 ## [0.2.0] - 2025-11-17
 
 ### Added